divide implementaion
